pyproject.toml
[project]
name = "winnie-the-pooh-qwen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets==3.5.0",
    "einops>=0.8.2",
    "flask>=3.1.2",
    "huggingface-hub==0.27.0",
    "jaxtyping>=0.3.6",
    "nnsight==0.3.7",
    "odps>=3.5.1",
    "pyahocorasick>=2.3.0",
    "pydantic>=2.12.5",
    "strenum>=0.4.15",
    "tokenizers==0.20.3",
    "tqdm>=4.67.1",
    "transformer-lens==2.15.4",
    "wandb>=0.24.1",
    "zstandard>=0.25.0",
]
